rename run-test to run-tests-lib
