+++ src/commands/synthetics/index.ts
@@ -2,6 +2,6 @@ export {CiError, CriticalError} from './errors'
 export * from './interfaces'
 export {DefaultReporter} from './reporters/default'
 export {JUnitReporter} from './reporters/junit'
-export {executeTests, execute} from './run-test'
+export {executeTests, execute} from './run-tests-lib'
 export {DEFAULT_COMMAND_CONFIG} from './command'
 export * as utils from './utils'
